+++ panzer/__init__.py
@@ -1,3 +1,5 @@
+from .credentials import CredentialManager
+from .exchanges.binance.client import BinanceClient
 from .exchanges.binance.public import BinancePublicClient
 
-__all__ = ["BinancePublicClient"]
+__all__ = ["BinanceClient", "BinancePublicClient", "CredentialManager"]